+++ multicore-sys-monitor@ccadeptic23/files/multicore-sys-monitor@ccadeptic23/3.2/applet.js
@@ -20,25 +20,29 @@
  * */
 var ImportError = false; //flag import error
 var ImportErrorMsg = "";
+const Applet = imports.ui.applet;
+const Lang = imports.lang;
+const Mainloop = imports.mainloop;
+const St = imports.gi.St;
+//const Cairo = imports.cairo;
+const Gio = imports.gi.Gio;
+const Gtk = imports.gi.Gtk;
+const GLib = imports.gi.GLib;
+const NMClient = imports.gi.NMClient;
+const NetworkManager = imports.gi.NetworkManager;
+const Cinnamon = imports.gi.Cinnamon;
+const Gettext = imports.gettext;
+
+const UUID = "multicore-sys-monitor@ccadeptic23";
+
+var SpawnProcess = null; //defined in main (my library)
+var ErrorApplet = null; //defined in main (my library)
+var ConfigSettings = null; //defined in main (my library)
+var Graphs = null; //defined in main (my library)
+var DataProviders = null; //defined in main (my library)
 
 try {
-  var Applet = imports.ui.applet;
-  var Lang = imports.lang;
-  var Mainloop = imports.mainloop;
-  var St = imports.gi.St;
-  var Gtk = imports.gi.Gtk;
-  var GLib = imports.gi.GLib;
-  var Cinnamon = imports.gi.Cinnamon;
-  var Gettext = imports.gettext;
-  var GTop = imports.gi.GTop; //psst this is really only to see if we can
-  var UUID = "multicore-sys-monitor@ccadeptic23";
-  var SpawnProcess = null; //defined in main (my library)
-  var ErrorApplet = null; //defined in main (my library)
-  var ConfigSettings = null; //defined in main (my library)
-  var Graphs = null; //defined in main (my library)
-  var DataProviders = null; //defined in main (my library)
-
-
+  const GTop = imports.gi.GTop; //psst this is really only to see if we can
 } catch (err) {
   ImportError = true;
   global.logError(err);
@@ -71,6 +75,12 @@ MyApplet.prototype = {
     this.configSettings = new ConfigSettings(this.configfilepath);
 
     this._initContextMenu();
+
+    if (St.Widget.get_default_direction() === St.TextDirection.RTL) {
+      this._applet_tooltip._tooltip.set_style('text-align: right; font-family: monospace;');
+    } else {
+      this._applet_tooltip._tooltip.set_style('text-align: left; font-family: monospace;');
+    }
 
     //this.gtop = new GTop.glibtop_cpu();
 
@@ -319,11 +329,22 @@ function main(metadata, orientation) {
   if (ImportError) {
     ErrorApplet = imports.ErrorApplet;
     var errmsg = ImportErrorMsg;
-    //we dont have the gtop package
-    if (typeof GTop === 'undefined') {
+    var myErrorApplet = new ErrorApplet.ErrorImportApplet(orientation, errmsg);
+    if (typeof GTop === 'undefined') { //we don't have the gtop package
       errmsg = _("Please install \"gir1.2-gtop-2.0\" package.");
+      let _is_apturl_present = GLib.find_program_in_path("apturl");
+      if (_is_apturl_present) {
+        const Extension = imports.ui.extension;
+        const PopupMenu = imports.ui.popupMenu;
+        let restart_button = new PopupMenu.PopupMenuItem(_("Reload this applet"));
+        restart_button.connect('activate', Lang.bind(this, function(event) {
+          Extension.reloadExtension(UUID, Extension.Type.APPLET);
+        }));
+        myErrorApplet.menu.addMenuItem(restart_button);
+        const Util = imports.misc.util;
+        Util.spawnCommandLine("apturl apt://gir1.2-gtop-2.0");
+      }
     }
-    let myErrorApplet = new ErrorApplet.ErrorImportApplet(orientation, errmsg);
     return myErrorApplet;
   } else {
     ConfigSettings = imports.ConfigSettings.ConfigSettings;